requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.11.4
pandas==2.2.3
sqlalchemy==2.0.40
websockets==13.1
